+++ src/app.py
@@ -1,22 +1,20 @@
+import os
+from typing import Annotated
 import logging
-from quart import Quart
-from quart_jwt_extended import JWTManager
-from upload import upload_bp
-from user import user_bp
+from fastapi import Depends, FastAPI
+from fastapi.security import OAuth2PasswordBearer
 
-app = Quart(__name__)
-app.config['JWT_SECRET_KEY'] = 'your_jwt_secret_key'  # Change this to a random secret key
-jwt = JWTManager(app)
+from .upload import upload_bp
+from .user import user_bp
+
+app = FastAPI()
+
 logging.getLogger().setLevel(logging.DEBUG)
-app.register_blueprint(upload_bp)
-app.register_blueprint(user_bp)
 
-@app.route('/')
+# Register blueprints (routers in FastAPI)
+app.include_router(upload_bp, prefix="/upload")
+app.include_router(user_bp, prefix="/user")
+
+@app.get("/")
 async def hello_world():
-    return 'Hello, World!'
-
-def main():
-    app.run(host="0.0.0.0", port=5000, debug=True)
-
-if __name__ == '__main__':
-    main()
+    return {"message": "Hello World!"}